Eliminate duplicated constants and text_with_caret per OCR review for #178
- Extract text_with_caret to shared src/ui/util.rs, eliminating three
  identical copies (form_content.rs, new_agent.rs, new_repository.rs)
- Export HELP_CHROME_ROWS, HELP_MODAL_WIDTH, and HELP_TITLE from
  src/ui/modals/help.rs as the single source of truth; geometry.rs and
  content.rs import these instead of maintaining their own copies
- Add explanatory comment for Search variant in active_overlay_for
  (it is an in-band filter, not a blocking overlay)
