+++ src/mouse_routing.rs
@@ -185,11 +185,13 @@ fn active_overlay_for(state: &AppState) -> jefe::selection::OverlayPane {
         }
         // Explicit match (not wildcard) so new ModalState variants force a
         // conscious overlay-routing decision here (issue #178 z-order).
+        jefe::state::ModalState::None
+        // Search is an in-band filter mode (SplitScreen's filter bar), not a
+        // blocking overlay — the base screen remains visible and interactive.
+        | jefe::state::ModalState::Search { .. }
         // ThemePicker renders as a full-screen modal but does not yet have a
         // content projection — no selection, but no PTY forwarding either
         // (pre-existing behavior; the base screen is not visible behind it).
-        jefe::state::ModalState::None
-        | jefe::state::ModalState::Search { .. }
         | jefe::state::ModalState::ThemePicker { .. } => {}
     }
     // Positioned overlays (choosers) — checked only when no full-screen modal.
